feat: 系统升级支持数据库自动迁移，避免手动执行升级脚本

+++ server/internal/sys/api/form/resource.go
@@ -1,7 +1,7 @@
 package form
 
 type ResourceForm struct {
-	Pid    int            `json:"pid"`
+	Pid    int64          `json:"pid"`
 	Id     int            `json:"id"`
 	Code   string         `json:"code" binding:"required"`
 	Name   string         `json:"name" binding:"required"`
